externalcommand: remove deprecated old class
