--- joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/utils/ExternalCommand.scala
@@ -1,6 +1,0 @@
-package io.joern.x2cpg.utils
-
-import io.shiftleft.semanticcpg.utils.ExternalCommandImpl
-
-@Deprecated(since = "Use `io.shiftleft.semanticcpg.utils.ExternalCommand` instead")
-object ExternalCommand extends ExternalCommandImpl
